+++ .bumpversion.cfg
@@ -1,8 +1,8 @@
 [bumpversion]
-current_version = 0.3.3.post2
+current_version = 2.0.0rc1
 commit = False
 tag = False
 
-[bumpversion:file:setup.py]
-search = __version__ = "{current_version}"
-replace = __version__ = "{new_version}"
+[bumpversion:file:pyproject.toml]
+search = "{current_version}"
+replace = "{new_version}"
