+++ .idea/workspace.xml
@@ -5,10 +5,27 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="3dc13afa-14b0-4d64-a848-dea06aab792d" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/pyproject.toml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.bumpversion.cfg" beforeDir="false" afterPath="$PROJECT_DIR$/.bumpversion.cfg" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/m2r2/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r2/__init__.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/m2r2/m2r2.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r2/m2r2.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/m2r2/sphinx_m2r2.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r2/sphinx_m2r2.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/README.md" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/docs/conf.py" beforeDir="false" afterPath="$PROJECT_DIR$/docs/conf.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/cli/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/cli/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/cli/m2r2.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/cli/m2r2.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/constants.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/constants.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/directives.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/directives.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/m2r2.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/m2r2.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/parser.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/parser.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/rst_parser.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/rst_parser.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/rst_renderer.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/rst_renderer.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/sphinx_m2r2.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/sphinx_m2r2.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/m2r2/typing.py" beforeDir="false" afterPath="$PROJECT_DIR$/m2r3/typing.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/requirements-dev.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/requirements-test.txt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/setup.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/tests/test_cli.py" beforeDir="false" afterPath="$PROJECT_DIR$/tests/test_cli.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/tests/test_renderer.py" beforeDir="false" afterPath="$PROJECT_DIR$/tests/test_renderer.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -29,21 +46,24 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent">{
-  &quot;keyToString&quot;: {
-    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
-    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
-    &quot;git-widget-placeholder&quot;: &quot;v2&quot;,
-    &quot;last_opened_file_path&quot;: &quot;/home/daquintero/flexcompute/m2r2&quot;,
-    &quot;node.js.detected.package.eslint&quot;: &quot;true&quot;,
-    &quot;node.js.detected.package.tslint&quot;: &quot;true&quot;,
-    &quot;node.js.selected.package.eslint&quot;: &quot;(autodetect)&quot;,
-    &quot;node.js.selected.package.tslint&quot;: &quot;(autodetect)&quot;,
-    &quot;nodejs_package_manager_path&quot;: &quot;npm&quot;,
-    &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "git-widget-placeholder": "poetry__packaging",
+    "last_opened_file_path": "/home/daquintero/flexcompute/m2r3",
+    "node.js.detected.package.eslint": "true",
+    "node.js.detected.package.tslint": "true",
+    "node.js.selected.package.eslint": "(autodetect)",
+    "node.js.selected.package.tslint": "(autodetect)",
+    "nodejs_package_manager_path": "npm",
+    "vue.rearranger.settings.migration": "true"
   }
-}</component>
+}]]></component>
   <component name="RecentsManager">
+    <key name="CopyFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$" />
+    </key>
     <key name="MoveFile.RECENT_KEYS">
       <recent name="$PROJECT_DIR$/m2r2" />
     </key>
@@ -65,6 +85,7 @@
       <updated>1703681877093</updated>
       <workItem from="1703681878094" duration="7005000" />
       <workItem from="1704276080142" duration="10450000" />
+      <workItem from="1705052475859" duration="2820000" />
     </task>
     <servers />
   </component>
