+++ pyproject.toml
@@ -1,0 +1,87 @@
+[tool.poetry]
+name = "m2r3"
+version = "2.0.0rc1"
+description = "Markdown and reStructuredText in a single file."
+authors = [
+    "Hiroyuki Takagi <miyako.dev@gmail.com>",
+    "CrossNox <ijmermet+m2r2@gmail.com>",
+    "daquintero <dario a quintero at gmail dot com>"
+]
+homepage = "http://www.sphinx-doc.org/"
+license = "MIT" # Replace with your project's license, if different
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Environment :: Console",
+    "Environment :: Web Environment",
+    "Framework :: Sphinx :: Extension",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: BSD License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Topic :: Text Processing",
+    "Topic :: Documentation",
+    "Topic :: Documentation :: Sphinx",
+    "Topic :: Utilities"
+]
+
+[tool.poetry.scripts]
+m2r2 = "m2r2:main"
+
+[tool.poetry.dependencies]
+python = ">=3.9"
+mistune = ">=2.0.0,<3.0"
+docutils = ">=0.16,<1.0"
+
+# Testing and linting
+black = {version="22.3.0", optional=true}
+bump2version = {version="1.0.1", optional=true}
+flake8 = {version="3.7.8", optional=true}
+isort = {version="4.3.21", optional=true}
+livereload = {version="*", optional=true}
+mypy = {version="0.770", optional=true}
+nox = {version="*", optional=true}
+pre-commit = {version="2.2.0", optional=true}
+pygments = {version="*", optional=true}
+pylint = {version="2.4.4", optional=true}
+pytest = {version="*", optional=true}
+pytest-cov = {version="*", optional=true}
+six = {version="*", optional=true}
+sphinx = {version="*", optional=true}
+
+[tool.poetry.extras]
+develop = [
+    "black",
+    "bump2version",
+    "flake8",
+    "isort",
+    "livereload",
+    "mypy",
+    "nox",
+    "pre-commit",
+    "pylint",
+    "pytest",
+    "pytest-cov",
+    "six",
+    "sphinx"
+]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.mypy]
+python_version = "3.8"
+show_column_numbers = true
+show_error_context = true
+ignore_missing_imports = true
+follow_imports = "skip"
+incremental = true
+check_untyped_defs = true
+warn_unused_ignores = true
+
+[tool.flake8]
+show-source = true
+builtins = "unicode"
