--- requirements-dev.txt
@@ -1,10 +1,0 @@
--r requirements-test.txt
-black==22.3.0
-bump2version==1.0.1
-flake8==3.7.8
-isort==4.3.21
-livereload
-mypy==0.770
-nox
-pre-commit==2.2.0
-pylint==2.4.4
