--- requirements-test.txt
@@ -1,4 +1,0 @@
-pytest
-pytest-cov
-six
-sphinx
